// src/app/_shared/models/user-details.ts
export class UserDetails {
    userId: number = 0;
    username: string = '';
    firstName: string = '';
    middleName: string = '';
    lastName: string = '';
    email: string = '';
    phoneNo: string = '';
    branchId: number | null = 0;
    branchName: string = '';
    roleId: number | null = 0;
    roleName: string = '';
}